+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added SonarQube analysis
 - Track alcohol-free beer separately in drink metadata and statistics.
 
+### Fixed
+- Fix logout without push token
+
 ## [1.2.1] - 2026-04-13
 ### Fixed
 - Fix failing tests in CI pipeline